microservico-embed/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
requests
